# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP-сервер на Go с автоматической генерацией Swagger при старте",
        "title": "Configurator",
        "contact": {
            "name": "Lorgar Avrelian",
            "url": "https://github.com/Lorgar-Avrelian",
            "email": "victor-14-244@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
...
        "/api/v1/thresholds": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "11. Конфигурация: Пороги"
                ],
                "summary": "Получить все пороги",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Threshold"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "11. Конфигурация: Пороги"
                ],
                "summary": "Создать порог",
                "parameters": [
                    {
                        "description": "Данные порога",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ThresholdCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Возвращает раскрытую цепочку условий",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Threshold"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/thresholds/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "11. Конфигурация: Пороги"
                ],
                "summary": "Получить порог по ID вместе со всей цепочкой родителей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Порога",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Threshold"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "11. Конфигурация: Пороги"
                ],
                "summary": "Обновить порог по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Порога",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ThresholdUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Threshold"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "11. Конфигурация: Пороги"
                ],
                "summary": "Удалить порог по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Порога",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "configurator_internal_dto.ThresholdCreate": {
            "type": "object",
            "required": [
                "level",
                "operator",
                "prev_operator",
                "source_internal_order",
                "source_model",
                "source_param",
                "type",
                "value"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "level": {
                    "type": "string",
                    "example": "WARNING"
                },
                "operator": {
                    "type": "string",
                    "example": "=="
                },
                "prev_operator": {
                    "type": "string",
                    "example": "\u0026\u0026"
                },
                "previous_id": {
                    "type": "integer",
                    "example": 2
                },
                "source_internal_order": {
                    "type": "integer",
                    "example": 1
                },
                "source_model": {
                    "type": "integer",
                    "example": 10
                },
                "source_param": {
                    "type": "string",
                    "example": "temperature"
                },
                "target_device": {
                    "type": "integer",
                    "example": 5
                },
                "target_param": {
                    "type": "string",
                    "example": "alarm_state"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "value": {
                    "type": "string",
                    "example": "85"
                }
            }
        },
        "configurator_internal_dto.ThresholdUpdate": {
            "type": "object",
            "required": [
                "level",
                "operator",
                "prev_operator",
                "source_internal_order",
                "source_model",
                "source_param",
                "type",
                "value"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "level": {
                    "type": "string",
                    "example": "ALARM"
                },
                "operator": {
                    "type": "string",
                    "example": "\u003e="
                },
                "prev_operator": {
                    "type": "string",
                    "example": "\u0026\u0026"
                },
                "previous_id": {
                    "type": "integer",
                    "example": 2
                },
                "source_internal_order": {
                    "type": "integer",
                    "example": 1
                },
                "source_model": {
                    "type": "integer",
                    "example": 10
                },
                "source_param": {
                    "type": "string",
                    "example": "temperature"
                },
                "target_device": {
                    "type": "integer",
                    "example": 5
                },
                "target_param": {
                    "type": "string",
                    "example": "alarm_state"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "value": {
                    "type": "string",
                    "example": "90"
                }
            }
        },
...
        "configurator_internal_model.Threshold": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "level": {
                    "type": "string",
                    "example": "WARNING"
                },
                "operator": {
                    "type": "string",
                    "example": "=="
                },
                "prev_operator": {
                    "type": "string",
                    "example": "\u0026\u0026"
                },
                "previous_id": {
                    "type": "integer",
                    "example": 2
                },
                "previous_threshold": {
                    "$ref": "#/definitions/configurator_internal_model.Threshold"
                },
                "source_internal_order": {
                    "type": "integer",
                    "example": 1
                },
                "source_model": {
                    "type": "integer",
                    "example": 10
                },
                "source_param": {
                    "type": "string",
                    "example": "temperature"
                },
                "target_device": {
                    "type": "integer",
                    "example": 5
                },
                "target_param": {
                    "type": "string",
                    "example": "alarm_state"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "value": {
                    "type": "string",
                    "example": "85"
                }
            }
        }
    },
...
}